+++ utils.c
@@ -26,7 +26,7 @@ void pprintf(const char* fmt, ...) {
 
 // FYI: allocates memory that consumer is responsible for
 char* qstrcopy(char* src) {
-  int len = strlen(src) + 1;
+  size_t len = strlen(src) + 1;
   char* dest = mmalloc(len);
   memset(dest, '\0', len);
   return strcpy(dest, src);
